security-controls-ingestion/requirements.txt
google-cloud-firestore==2.13.1
python-dotenv==1.0.0
google-cloud-asset==3.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-cloud-securitycentermanagement==1.0.0
